Refactor and modularize plugin code for reusability and local development
- Move and rename core DigitalOcean Spaces and Maven artifacts classes to local packages for better separation
- Add ProjectSync and JarExplorer plugins with corresponding tasks for project file synchronization and jar exploration
- Introduce DigitalOceanSpacesPublishPluginService for cleaner plugin application and task registration
- Update build scripts to register new plugins and manage versions via project properties
- Enhance extension classes and adapters for improved configuration and logging
- Refactor UploadToDigitalOceanSpacesService to use injected version check service
- Clean up and comment build scripts for clarity and future extensibility
