+++ buildSrc/src/main/kotlin/io/violabs/plugins/local/publishing/digitalocean/adapter/S3BuilderAdapter.kt
@@ -1,0 +1,12 @@
+package io.violabs.plugins.local.publishing.digitalocean.adapter
+
+import software.amazon.awssdk.services.s3.S3Client
+
+interface S3BuilderAdapter {
+    val accessKey: String
+    val secretKey: String
+    val endpoint: String
+    val region: String
+
+    fun build(): S3Client
+}
