+++ buildSrc/src/main/kotlin/io/violabs/plugins/local/publishing/digitalocean/client/DefaultDigitalOceanSpacesClient.kt
@@ -1,8 +1,8 @@
-package io.violabs.plugins.open.publishing.digitalocean.client
+package io.violabs.plugins.local.publishing.digitalocean.client
 
-import io.violabs.plugins.open.publishing.digitalocean.adapter.DefaultS3BuilderAdapter
-import io.violabs.plugins.open.publishing.digitalocean.adapter.S3BuilderAdapter
-import io.violabs.plugins.open.publishing.digitalocean.domain.DigitalOceanSpacesExtension
+import io.violabs.plugins.local.publishing.digitalocean.adapter.DefaultS3BuilderAdapter
+import io.violabs.plugins.local.publishing.digitalocean.adapter.S3BuilderAdapter
+import io.violabs.plugins.local.publishing.digitalocean.domain.DigitalOceanSpacesExtension
 import org.gradle.api.logging.Logger
 import software.amazon.awssdk.services.s3.S3Client
 import software.amazon.awssdk.services.s3.model.ObjectCannedACL
