+++ buildSrc/src/main/kotlin/io/violabs/plugins/local/publishing/digitalocean/domain/DigitalOceanSpacesExtension.kt
@@ -1,4 +1,4 @@
-package io.violabs.plugins.open.publishing.digitalocean.domain
+package io.violabs.plugins.local.publishing.digitalocean.domain
 
 /**
  * Extension for configuring the plugin
