+++ buildSrc/src/main/kotlin/io/violabs/plugins/local/publishing/digitalocean/task/DigitalOceanSpacesUploadTask.kt
@@ -1,9 +1,9 @@
-package io.violabs.plugins.open.publishing.digitalocean.task
+package io.violabs.plugins.local.publishing.digitalocean.task
 
-import io.violabs.plugins.open.publishing.digitalocean.adapter.DefaultProjectAdapter
-import io.violabs.plugins.open.publishing.digitalocean.client.DigitalOceanSpacesClient
-import io.violabs.plugins.open.publishing.digitalocean.service.CheckVersionDigitalOceanSpacesService
-import io.violabs.plugins.open.publishing.digitalocean.service.UploadToDigitalOceanSpacesService
+import io.violabs.plugins.local.publishing.digitalocean.adapter.DefaultProjectAdapter
+import io.violabs.plugins.local.publishing.digitalocean.client.DigitalOceanSpacesClient
+import io.violabs.plugins.local.publishing.digitalocean.service.CheckVersionDigitalOceanSpacesService
+import io.violabs.plugins.local.publishing.digitalocean.service.UploadToDigitalOceanSpacesService
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Internal
