--- publishing/digital-ocean-spaces-core/src/main/kotlin/io/violabs/plugins/open/publishing/digitalocean/core/adapter/S3BuilderAdapter.kt
@@ -1,12 +1,0 @@
-package io.violabs.plugins.open.publishing.digitalocean.core.adapter
-
-import software.amazon.awssdk.services.s3.S3Client
-
-interface S3BuilderAdapter {
-    val accessKey: String
-    val secretKey: String
-    val endpoint: String
-    val region: String
-
-    fun build(): S3Client
-}
