+++ publishing/maven-generated-artifacts/build.gradle.kts
@@ -1,13 +1,17 @@
 import io.violabs.plugins.local.secrets.getPropertyOrEnv
 
+val publishingMavenGeneratedArtifactsVersion: String by rootProject.extra
+
 plugins {
     `kotlin-dsl`
-    id("io.violabs.plugins.local.publishing.digital-ocean-spaces")
     id("org.jetbrains.dokka")
+    id("io.violabs.plugins.local.publishing.project-sync")
+    id("io.violabs.plugins.local.publishing.maven-generated-artifacts")
+    id("io.violabs.plugins.local.publishing.digital-ocean-spaces")
 }
 
 group = "io.violabs.plugins.open.publishing"
-version = "0.0.2"
+version = publishingMavenGeneratedArtifactsVersion
 
 buildscript {
     repositories {
@@ -27,12 +31,22 @@ repositories {
     mavenCentral()
 }
 
+projectSync {
+    val projectFile = rootProject.layout
+        .projectDirectory
+        .asFile
+    syncSource = projectFile
+        .resolve("publishing/maven-generated-artifacts/src/main/kotlin/io/violabs/plugins/open/publishing/mavengenerated")
+    syncTarget = projectFile
+        .resolve("buildSrc/src/main/kotlin/io/violabs/plugins/local/publishing/mavengenerated")
+}
+
 gradlePlugin {
     plugins {
-        create("mavenGeneratedArtifacts") {
+        create("mavenGeneratedArtifactsPlugin") {
             id = "io.violabs.plugins.open.publishing.maven-generated-artifacts"
             version = project.version.toString()
-            implementationClass = "io.violabs.plugins.open.publishing.ManualMavenArtifactsPlugin"
+            implementationClass = "io.violabs.plugins.open.publishing.MavenGeneratedArtifactsPlugin"
         }
     }
 }
