+++ publishing/maven-generated-artifacts/src/main/kotlin/io/violabs/plugins/open/publishing/mavengenerated/MavenGeneratedArtifactsPlugin.kt
@@ -1,0 +1,13 @@
+package io.violabs.plugins.open.publishing.mavengenerated
+
+import io.violabs.plugins.open.publishing.mavengenerated.service.MavenGeneratedArtifactsPublishPluginService
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+class MavenGeneratedArtifactsPlugin : Plugin<Project> {
+    override fun apply(project: Project) {
+        val service = MavenGeneratedArtifactsPublishPluginService()
+
+        service.apply(project)
+    }
+}
